requirements.txt
Flask==1.1.1
gunicorn==19.9.0
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.17.2
scipy>=0.15.1
scikit-learn==0.23.1
matplotlib>=1.4.3
pandas>=0.19
jsonify
requests
tensorflow-cpu
certifi==2019.11.28
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.0
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
Werkzeug==0.16.0
wincertstore==0.2
